time

+++ microservices/calculation_unit/src/websocket_format.rs
@@ -3,19 +3,19 @@ use serde::{Serialize, Deserialize};
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct ClientInfo {
-	sent_time: usize,
+	sent_time: f64,
 }
 #[derive(Serialize, Deserialize, Debug)]
 pub struct AuthproxyInfo {
-	sent_time: usize,
+	sent_time: f64,
 }
 #[derive(Serialize, Deserialize, Debug)]
 pub struct RequestSyncInfo {
-	sent_time: usize,
+	sent_time: f64,
 }
 #[derive(Serialize, Deserialize, Debug)]
 pub struct CalculationUnitInfo {
-	sent_time: usize,
+	sent_time: f64,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
